+++ constants/pagePath.ts
@@ -2,6 +2,6 @@ export const PAGE_PATH = {
   MAIN: "/",
   SIGN_IN: "/sign-in",
   DASHBOARD: "/dashboard",
-  CALENDER: "/calender",
+  CALENDAR: "/calendar",
   GROUP: "/group",
 } as const;
